+++ server/demo/src/main/java/example/domain/comment/service/CommentService.java
@@ -1,5 +1,6 @@
 package example.domain.comment.service;
 
+import example.domain.artistPost.entity.ArtistPost;
 import example.domain.comment.entity.Comment;
 import example.domain.comment.repository.CommentRepository;
 import example.domain.feedPost.entity.FeedPost;
@@ -26,6 +27,7 @@ public class CommentService {
     < 댓글 등록>
      */
     public Comment createComment(Comment comment) {
+
         return commentRepository.save(comment);
     }
 
@@ -57,8 +59,8 @@ public class CommentService {
     }
 
 
-    public Comment updateComment(Comment comment){
-        Comment findComment = findVerifiedComment(comment.getId());
+    public Comment updateFanComment(int commentId, Comment comment){
+        Comment findComment = findVerifiedComment(commentId);
         if(comment.getFans().getId() != findComment.getFans().getId()) {
             throw new BusinessLogicException(ExceptionCode.COMMENT_AUTHOR_NOT_MATCH);
         }else{
@@ -73,8 +75,35 @@ public class CommentService {
         }
     }
 
-    public void deleteComment(int commentId){
+    public Comment updateArtistComment(int commentId, Comment comment){
         Comment findComment = findVerifiedComment(commentId);
+        if(comment.getArtist().getId() != findComment.getArtist().getId()) {
+            throw new BusinessLogicException(ExceptionCode.COMMENT_AUTHOR_NOT_MATCH);
+        }else{
+            Optional.ofNullable(comment.getContent()) // 내용 수정
+                    .ifPresent(commentContent -> findComment.setContent(commentContent));
+            Optional.ofNullable(comment.getFeedPost())
+                    .ifPresent(commentFeedPost -> findComment.setFeedPost(commentFeedPost));
+            Optional.ofNullable(comment.getCreatedAt())
+                    .ifPresent(commentCreatedAt -> findComment.setCreatedAt(commentCreatedAt)); // 업데이트 날짜 수정
+
+            return commentRepository.save(findComment);
+        }
+    }
+
+    public void deleteFeedPostComment(FeedPost feedPost, int commentId){
+        Comment findComment = findVerifiedComment(commentId);
+        if(feedPost.getFans().getId() != findComment.getFans().getId() || feedPost.getArtist().getId() != findComment.getArtist().getId()) {
+            throw new BusinessLogicException(ExceptionCode.COMMENT_AUTHOR_NOT_MATCH);
+        }
+        commentRepository.delete(findComment);
+    }
+
+    public void deleteArtistPostComment(ArtistPost artistPost, int commentId){
+        Comment findComment = findVerifiedComment(commentId);
+        if(artistPost.getFans().getId() != findComment.getFans().getId() || artistPost.getArtist().getId() != findComment.getArtist().getId()) {
+            throw new BusinessLogicException(ExceptionCode.COMMENT_AUTHOR_NOT_MATCH);
+        }
         commentRepository.delete(findComment);
     }
